refactor: major refactor
-  game page - clean, optimize code
- Geme components - move the props to a separate file, clean code
