+++ src/components/Game/Keyboard/index.ts
@@ -1,1 +1,2 @@
 export * from './Keyboard';
+export * from './Keyboard.props';
