+++ src/components/Game/Racing/Racing.props.ts
@@ -1,0 +1,7 @@
+export interface RacingProps {
+  gameSpeed: number;
+  isGame: boolean;
+  lettersNum: number;
+  idx: number;
+  setIsEnded: (arg: boolean) => void,
+}
