+++ src/components/Game/Racing/index.ts
@@ -1,1 +1,2 @@
 export * from './Racing';
+export * from './Racing.props';
