+++ src/components/Game/TextWindow/index.ts
@@ -1,1 +1,2 @@
 export * from './TextWindow';
+export * from './TextWindow.props';
