+++ src/components/Game/TrafficLight/TrafficLight.props.ts
@@ -1,0 +1,7 @@
+export interface TrafficLightProps {
+  id: 'timer' | 'speed' | 'accuracy',
+  color: 'red' | 'green' | 'yellow',
+  textInfo: string | number,
+  textDesc: string,
+  unit: string,
+}
