+++ src/components/Game/TrafficLight/index.ts
@@ -1,1 +1,2 @@
 export * from './TrafficLight';
+export * from './TrafficLight.props';
