+++ src/components/Game/texts.ts
@@ -9,16 +9,16 @@ export const texts: { [key: string]: string[] } = {
   ru: [
     'Если жить только для себя, своими мелкими заботами о собственном благополучии, то от прожитого не останется и следа. Если же жить для других, то другие сберегут то, чему служил, чему отдавал силы.',
     'Надо иметь и принципы в жизни. Одно правило в жизни должно быть у каждого человека, в его цели жизни, в его принципах жизни, в его поведении: надо прожить жизнь с достоинством.',
-    'Ради достоинства жизни надо уметь отказываться от мелких удовольствий и немалых тоже… Уметь извиняться, признавать перед другими ошибку - лучше, чем врать.',
+    'Ради достоинства жизни надо уметь отказываться от мелких удовольствий и немалых тоже... Уметь извиняться, признавать перед другими ошибку - лучше, чем врать.',
     'Жизнь - прежде всего творчество, но это не значит, что каждый человек, чтобы жить, должен родиться художником, балериной или ученым. Можно творить просто добрую атмосферу вокруг себя.',
     'Надо уважать всякого человека, какой бы он ни был жалкий и смешной. Надо помнить, что во всяком человеке живёт тот же дух, какой и в нас.',
   ],
 };
 
 export const getRandomText = (keyboardLang?: string) => {
-  const maxI = texts[keyboardLang ? keyboardLang : 'en'].length;
-  const minI = 0;
-  const r = Math.floor(Math.random() * (maxI - minI) + minI); // excl. max
+  const max = texts[keyboardLang ? keyboardLang : 'en'].length;
+  const min = 0;
+  const r = Math.floor(Math.random() * (max - min) + min); // excl. max
   return texts[keyboardLang ? keyboardLang : 'en'][r];
 };
