+++ src/pages/create-game/create-game.tsx
@@ -1,13 +1,19 @@
+import { useEffect } from 'react';
 import { Game, Header, Footer } from '../../components';
 import styles from './create-game.module.scss';
 
 export default function CreateGamePage() {
 
-  window.addEventListener('keydown', (e: KeyboardEvent) => {
+  const blockingScroll = (e: KeyboardEvent) => {
     if (e.code === 'Space' && e.target === document.body) {
       e.preventDefault();
     }
-  });
+  };
+
+  useEffect(() => {
+    window.addEventListener('keydown', (e: KeyboardEvent) => blockingScroll(e));
+    return window.removeEventListener('keydown', (e: KeyboardEvent) => blockingScroll(e));
+  }, []);
 
   return (
     <>
